Move config models to `zabbix_auto_config.config`
