+++ tests/test_config.py
@@ -1,23 +1,33 @@
 from __future__ import annotations
 
 from pathlib import Path
+from typing import Any
+from typing import Optional
 from unittest.mock import patch
 
 import pytest
 import structlog
 import tomli
-import zabbix_auto_config.models as models
 from hypothesis import given
 from hypothesis import settings
 from hypothesis import strategies as st
 from inline_snapshot import snapshot
 from pydantic import ValidationError
+from zabbix_auto_config.config import DBSettings
+from zabbix_auto_config.config import DBTableSettings
+from zabbix_auto_config.config import FailureStrategy
+from zabbix_auto_config.config import LoggingSettings
+from zabbix_auto_config.config import LogLevel
+from zabbix_auto_config.config import Settings
+from zabbix_auto_config.config import SourceCollectorSettings
+from zabbix_auto_config.config import ZabbixSettings
+from zabbix_auto_config.config import ZacSettings
 from zabbix_auto_config.config import get_config
 from zabbix_auto_config.config import load_config
 
 
 def test_sample_config(sample_config: str):
-    models.Settings(**tomli.loads(sample_config))
+    Settings(**tomli.loads(sample_config))
 
 
 def test_config_extra_field(
@@ -25,7 +35,7 @@ def test_config_extra_field(
 ):
     config = tomli.loads(sample_config)
     config["foo"] = "bar"
-    models.Settings(**config)
+    Settings(**config)
     assert len(log_output.entries) == 1
     assert log_output.entries == snapshot(
         [{"event": "Settings: Got unknown config field 'foo'.", "log_level": "warning"}]
@@ -39,24 +49,24 @@ def test_config_extra_field_allowed(
     config["foo"] = "bar"
 
     # Allow extra fields for this test
-    original_extra = models.Settings.model_config["extra"]  # pyright: ignore[reportTypedDictNotRequiredAccess]
+    original_extra = Settings.model_config["extra"]  # pyright: ignore[reportTypedDictNotRequiredAccess]
     try:
-        models.Settings.model_config["extra"] = "allow"
-        models.Settings(**config)
+        Settings.model_config["extra"] = "allow"
+        Settings(**config)
         assert len(log_output.entries) == 0
     finally:
-        models.Settings.model_config["extra"] = original_extra
+        Settings.model_config["extra"] = original_extra
 
 
 def test_sourcecollectorsettings_defaults():
     # Default setting should be valid
-    settings = models.SourceCollectorSettings(
+    settings = SourceCollectorSettings(
         module_name="foo",
         update_interval=60,
     )
 
     # Default strategy should be to use exponential backoff
-    assert settings.failure_strategy == models.FailureStrategy.BACKOFF
+    assert settings.failure_strategy == FailureStrategy.BACKOFF
 
     # Snapshot of values
     assert settings.model_dump() == snapshot(
@@ -81,7 +91,7 @@ def test_sourcecollectorsettings_no_tolerance() -> None:
     the value of error_duration is set to a non-zero value to ensure that
     the error is not discarded when calling RollingErrorCounter.check().
     """
-    settings = models.SourceCollectorSettings(
+    settings = SourceCollectorSettings(
         module_name="foo",
         update_interval=60,
         error_tolerance=0,
@@ -93,7 +103,7 @@ def test_sourcecollectorsettings_no_tolerance() -> None:
 
 def test_sourcecollectorsettings_no_error_duration():
     # In order to not have an error_duration, error_tolerance must be 0 too
-    settings = models.SourceCollectorSettings(
+    settings = SourceCollectorSettings(
         module_name="foo",
         update_interval=60,
         error_duration=0,
@@ -102,7 +112,7 @@ def test_sourcecollectorsettings_no_error_duration():
     assert settings.error_duration == snapshot(9999)
 
     # With tolerance we get a default value
-    settings = models.SourceCollectorSettings(
+    settings = SourceCollectorSettings(
         module_name="foo",
         update_interval=60,
         error_duration=0,
@@ -122,7 +132,7 @@ def test_sourcecollectorsettings_no_error_duration_fuzz(
 ):
     """Test model with a variety of update intervals and error tolerances"""
     # We only check that instantiating the model does not raise an exception
-    models.SourceCollectorSettings(
+    SourceCollectorSettings(
         module_name="foo",
         update_interval=update_interval,
         error_tolerance=error_tolerance,
@@ -133,7 +143,7 @@ def test_sourcecollectorsettings_duration_too_short():
     # Error_duration should be greater or equal to the product of
     # error_tolerance and update_interval
     with pytest.raises(ValidationError) as exc_info:
-        models.SourceCollectorSettings(
+        SourceCollectorSettings(
             module_name="foo",
             update_interval=60,
             error_tolerance=5,
@@ -159,7 +169,7 @@ def test_sourcecollectorsettings_duration_too_short():
 def test_sourcecollectorsettings_duration_negative():
     # We should not be able to pass in negative values to error_duration
     with pytest.raises(ValidationError) as exc_info:
-        models.SourceCollectorSettings(
+        SourceCollectorSettings(
             module_name="foo",
             update_interval=60,
             error_tolerance=5,
@@ -334,3 +344,359 @@ def test_get_config_find_config(sample_config_path: Path, tmp_path: Path) -> Non
         # Test that the loaded config matches the sample config
         # (compare in dumped JSON mode to remove excluded fields)
         assert config.model_dump(mode="json") == sample_config.model_dump(mode="json")
+
+
+@pytest.mark.parametrize(
+    "inp,expect",
+    [
+        # Lower case values
+        ["notset", LogLevel.NOTSET],
+        ["debug", LogLevel.DEBUG],
+        ["info", LogLevel.INFO],
+        ["warn", LogLevel.WARNING],  # alias for warning
+        ["warning", LogLevel.WARNING],
+        ["error", LogLevel.ERROR],
+        ["fatal", LogLevel.CRITICAL],  # alias for critical
+        ["critical", LogLevel.CRITICAL],
+        # Upper case values
+        ["NOTSET", LogLevel.NOTSET],
+        ["DEBUG", LogLevel.DEBUG],
+        ["INFO", LogLevel.INFO],
+        ["WARN", LogLevel.WARNING],  # alias for warning
+        ["WARNING", LogLevel.WARNING],
+        ["ERROR", LogLevel.ERROR],
+        ["FATAL", LogLevel.CRITICAL],  # alias for critical
+        ["CRITICAL", LogLevel.CRITICAL],
+        # Numeric values
+        [0, LogLevel.NOTSET],
+        [10, LogLevel.DEBUG],
+        [20, LogLevel.INFO],
+        [30, LogLevel.WARNING],
+        [40, LogLevel.ERROR],
+        [50, LogLevel.CRITICAL],
+        # Numeric values as strings
+        ["0", LogLevel.NOTSET],
+        ["10", LogLevel.DEBUG],
+        ["20", LogLevel.INFO],
+        ["30", LogLevel.WARNING],
+        ["40", LogLevel.ERROR],
+        ["50", LogLevel.CRITICAL],
+        # Invalid values (uses default level)
+        ["invalid", LogLevel.ERROR],
+        [None, LogLevel.ERROR],
+    ],
+)
+def test_log_level(inp: Any, expect: LogLevel) -> None:
+    assert LogLevel(inp) == expect
+
+
+def test_logging_settings_log_level_serialize() -> None:
+    conf = LoggingSettings()
+    conf.file.path = Path("/path/to/logfile.log")  # bogus path
+
+    # Serialize to dict (in JSON mode):
+    conf_dict = conf.model_dump(mode="json")
+    assert conf_dict["level"] == "INFO"
+    assert conf_dict["console"]["level"] == "INFO"
+    assert conf_dict["file"]["level"] == "INFO"
+
+    # Serialize to JSON:
+    conf_json = conf.model_dump_json(indent=2)
+    assert conf_json == snapshot(
+        """\
+{
+  "console": {
+    "enabled": true,
+    "format": "text",
+    "level": "INFO"
+  },
+  "file": {
+    "enabled": true,
+    "format": "json",
+    "level": "INFO",
+    "path": "/path/to/logfile.log",
+    "rotate": true,
+    "max_size_mb": 50,
+    "max_logs": 5
+  },
+  "level": "INFO",
+  "use_mp_handler": false
+}\
+"""
+    )
+
+
+@pytest.mark.parametrize(
+    "timeout,expect",
+    [
+        (1, 1),
+        (60, 60),
+        (1234, 1234),
+        (0, None),
+        pytest.param(
+            -1,
+            None,
+            marks=pytest.mark.xfail(
+                reason="Timeout must be 0 or greater.",
+                strict=True,
+                raises=ValidationError,
+            ),
+            id="-1",
+        ),
+    ],
+)
+def test_zabbix_settings_timeout(timeout: int, expect: Optional[int]) -> None:
+    settings = ZabbixSettings(
+        map_dir="",
+        url="",
+        username="",
+        password="",
+        dryrun=False,
+        timeout=timeout,
+    )
+    assert settings.timeout == expect
+
+
+def test_failure_strategy_supports_error_tolerance() -> None:
+    """Test that only EXIT and DISABLE support error tolerance."""
+    for strategy in FailureStrategy:
+        if strategy in (FailureStrategy.EXIT, FailureStrategy.DISABLE):
+            assert strategy.supports_error_tolerance()
+        else:
+            assert not strategy.supports_error_tolerance()
+
+
+def _get_zac_settings(config: Settings, db_uri: str) -> ZacSettings:
+    return ZacSettings(
+        source_collector_dir=config.zac.source_collector_dir,
+        host_modifier_dir=config.zac.host_modifier_dir,
+        db_uri=db_uri,
+        # Omit DBSettings
+    )
+
+
+def test_zacsettings_db_uri_all(config: Settings):
+    """Parse a PostgreSQL connection string with all args."""
+    settings = _get_zac_settings(
+        config,
+        "dbname='zac' user='zabbix' host='localhost' password='secret' port=5432 connect_timeout=2",
+    )
+
+    assert settings.db.model_dump(mode="json") == snapshot(
+        {
+            "user": "zabbix",
+            "password": "secret",
+            "dbname": "zac",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 2,
+            "tables": {"hosts": "hosts", "hosts_source": "hosts_source"},
+            "init": {"db": True, "tables": True},
+        }
+    )
+
+    assert settings.db.get_connect_kwargs() == snapshot(
+        {
+            "dbname": "zac",
+            "user": "zabbix",
+            "password": "secret",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 2,
+        }
+    )
+
+
+def test_zacsettings_db_uri_only_required(config: Settings):
+    """Parse a PostgreSQL connection string with only required args."""
+    settings = _get_zac_settings(
+        config,
+        "user='zabbix' password='secret'",
+    )
+
+    assert settings.db.model_dump(mode="json") == snapshot(
+        {
+            "user": "zabbix",
+            "password": "secret",
+            "dbname": "zac",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 5,
+            "tables": {"hosts": "hosts", "hosts_source": "hosts_source"},
+            "init": {"db": True, "tables": True},
+        }
+    )
+
+    assert settings.db.get_connect_kwargs() == snapshot(
+        {
+            "dbname": "zac",
+            "user": "zabbix",
+            "password": "secret",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 5,
+        }
+    )
+
+
+def test_zacsettings_db_uri_extra(config: Settings):
+    """Parse a PostgreSQL connection string with all args + extra args."""
+    settings = _get_zac_settings(
+        config,
+        "dbname='zac' user='zabbix' host='localhost' password='secret' port=5432 connect_timeout=2",
+    )
+
+    assert settings.db.model_dump(mode="json") == snapshot(
+        {
+            "user": "zabbix",
+            "password": "secret",
+            "dbname": "zac",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 2,
+            "tables": {"hosts": "hosts", "hosts_source": "hosts_source"},
+            "init": {"db": True, "tables": True},
+        }
+    )
+
+    assert settings.db.get_connect_kwargs() == snapshot(
+        {
+            "dbname": "zac",
+            "user": "zabbix",
+            "password": "secret",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 2,
+        }
+    )
+
+
+def test_zacsettings_db_uri_empty_values_and_extras(config: Settings):
+    """Parse a PostgreSQL connection string with some empty values and extra kwargs."""
+    settings = _get_zac_settings(
+        config,
+        "user='zabbix' password='' sslmode='require' passfile='/path/to/passfile'",
+    )
+
+    assert settings.db.model_dump(mode="json") == snapshot(
+        {
+            "user": "zabbix",
+            "password": "",
+            "dbname": "zac",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 5,
+            "tables": {"hosts": "hosts", "hosts_source": "hosts_source"},
+            "init": {"db": True, "tables": True},
+            "sslmode": "require",
+            "passfile": "/path/to/passfile",
+        }
+    )
+
+    assert settings.db.get_connect_kwargs() == snapshot(
+        {
+            "dbname": "zac",
+            "user": "zabbix",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 5,
+            "sslmode": "require",
+            "passfile": "/path/to/passfile",
+        }
+    )
+
+
+def test_zacsettings_db_uri_missing_all(config: Settings):
+    """Parse a PostgreSQL connection string with no args."""
+    settings = _get_zac_settings(
+        config,
+        "",
+    )
+
+    assert settings.db.model_dump(mode="json") == snapshot(
+        {
+            "user": "",
+            "password": "",
+            "dbname": "zac",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 2,
+            "tables": {"hosts": "hosts", "hosts_source": "hosts_source"},
+            "init": {"db": True, "tables": True},
+        }
+    )
+
+    assert settings.db.get_connect_kwargs() == snapshot(
+        {"dbname": "zac", "host": "localhost", "port": 5432, "connect_timeout": 2}
+    )
+
+
+def test_zacsettings_dbsettings(config: Settings) -> None:
+    """Test ZacSettings with DBSettings."""
+    config.zac.db = DBSettings(
+        user="zac",
+        password="secret",
+        dbname="zac",
+        host="localhost",
+        port=5432,
+        connect_timeout=2,
+        # Extra kwargs
+        sslmode="require",
+        passfile="/path/to/passfile",
+    )
+
+    assert config.zac.db.get_connect_kwargs() == snapshot(
+        {
+            "dbname": "zac",
+            "user": "zac",
+            "password": "secret",
+            "host": "localhost",
+            "port": 5432,
+            "connect_timeout": 2,
+            "sslmode": "require",
+            "passfile": "/path/to/passfile",
+        }
+    )
+
+
+def test_dbsettings_extra_kwargs() -> None:
+    """Test DBSettings with extra kwargs."""
+    db = DBSettings(
+        user="zac",
+        password="secret",
+        dbname="zac",
+        host="localhost",
+        port=5432,
+        connect_timeout=2,
+        # Extra kwargs (included)
+        sslmode="require",
+        passfile="/path/to/passfile",
+        # Extra kwargs (ignored)
+        dict_kwarg={"extra": "value"},
+        list_kwarg=["extra", "value"],
+        tuple_kwarg=("extra", "value"),
+        set_kwarg={"extra", "value"},
+        none_kwarg=None,
+    )
+
+    assert db.extra_kwargs() == snapshot(
+        {"sslmode": "require", "passfile": "/path/to/passfile"}
+    )
+
+
+def test_dbtablesettings_duplicate_names() -> None:
+    """Test DBTableSettings with duplicate names."""
+    with pytest.raises(ValueError, match="Duplicate table name: 'hosts'"):
+        DBTableSettings(
+            hosts="hosts",
+            hosts_source="hosts",
+        )
+
+
+def test_dbtablesettings_empty_name() -> None:
+    """Test DBTableSettings with empty name."""
+    with pytest.raises(ValueError, match="empty"):
+        DBTableSettings(
+            hosts="",
+            hosts_source="hosts_source",
+        )
